Stop build_fixtures.py pretending to build fixtures
The name was a trap. contracts/build_fixtures.py wrote adapters.json,
config.json, history.json and research-result.*.json — but its own docstring
admitted it was legacy, that it rebuilt the retired Airbnb-themed demo, and
that it could not run at all without the gitignored data/ directory. Anyone who
ran the thing the name told them to run would have clobbered the current
YouTube-growth fixtures with old demo content.

Meanwhile the fixtures it claimed to own come from two other places entirely:
build_mock_fixtures.py generates the research-result and history ones from the
curated set in contracts/mock_videos/, and adapters.json, config.json and
progress-events.jsonl are hand-maintained. Nothing said so.

It could not just be deleted: test_pipeline.py imports the module and diffs its
normalize_video against the backend's, so the two independent implementations
have to keep agreeing. That normalizer is the only part worth keeping.

So strip it to the normalizer (478 lines to 101), rename it normalize_reference
to say what it is, and drop the duplicated adapter model list it carried as a
bonus third copy. Rewrite contracts/README.md to spell out which fixtures are
generated and which are hand-maintained, since that was the underlying
confusion. BUILD_PLAN.md keeps its original W0.2 claim — it was true then —
with a note that it has since been superseded.
